requirements.txt
flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.38
langchain-huggingface==0.0.3
langchain-openai==0.1.23
langchain-ollama==0.1.3
faiss-cpu==1.8.0.post1
huggingface-hub==0.24.6
numpy==1.26.4

